requirements.txt
Flask-Mail==0.9.1
Flask-CORS==4.0.1
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1